Tidy up `Startup`
The Startup class is getting a bit unwieldy. Update to use extension/static classes to help keep things organised.
