+++ GetIntoTeachingApi/Program.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using GetIntoTeachingApi.AppStart;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
@@ -15,10 +16,7 @@ namespace GetIntoTeachingApi
 
         public static IHostBuilder CreateHostBuilder(string[] args) =>
             Host.CreateDefaultBuilder(args)
-                .UseDefaultServiceProvider((_, options) =>
-                {
-                    options.ValidateScopes = false;
-                })
+                .UseDefaultServiceProvider((_, options) => options.ValidateScopes = false)
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
                     webBuilder.UseSentry();
