+++ GetIntoTeachingApiTests/Integration/RateLimitTests.cs
@@ -4,7 +4,7 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using FluentAssertions;
-using GetIntoTeachingApi;
+using GetIntoTeachingApi.AppStart;
 using GetIntoTeachingApiTests.Helpers;
 using Xunit;
